Turns out size_t and int discrepancies are REALLY IMPORTANT to zend_parse_arguments

+++ ext/crt.c
@@ -151,7 +151,7 @@ PHP_FUNCTION(aws_crt_credentials_options_release) {
 PHP_FUNCTION(aws_crt_credentials_options_set_access_key_id) {
     zend_ulong php_options = 0;
     const char *access_key_id = NULL;
-    int access_key_id_len = 0;
+    size_t access_key_id_len = 0;
 
     if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ls", &php_options, &access_key_id, &access_key_id_len) ==
         FAILURE) {
@@ -159,13 +159,13 @@ PHP_FUNCTION(aws_crt_credentials_options_set_access_key_id) {
     }
 
     aws_crt_credentials_options *options = (void *)php_options;
-    aws_crt_credentials_options_set_access_key_id(options, (uint8_t *)access_key_id, (size_t)access_key_id_len);
+    aws_crt_credentials_options_set_access_key_id(options, (uint8_t *)access_key_id, access_key_id_len);
 }
 
 PHP_FUNCTION(aws_crt_credentials_options_set_secret_access_key) {
     zend_ulong php_options = 0;
     const char *secret_access_key = NULL;
-    int secret_access_key_len = 0;
+    size_t secret_access_key_len = 0;
 
     if (zend_parse_parameters(
             ZEND_NUM_ARGS() TSRMLS_CC, "ls", &php_options, &secret_access_key, &secret_access_key_len) == FAILURE) {
@@ -173,14 +173,13 @@ PHP_FUNCTION(aws_crt_credentials_options_set_secret_access_key) {
     }
 
     aws_crt_credentials_options *options = (void *)php_options;
-    aws_crt_credentials_options_set_secret_access_key(
-        options, (uint8_t *)secret_access_key, (size_t)secret_access_key_len);
+    aws_crt_credentials_options_set_secret_access_key(options, (uint8_t *)secret_access_key, secret_access_key_len);
 }
 
 PHP_FUNCTION(aws_crt_credentials_options_set_session_token) {
     zend_ulong php_options = 0;
     const char *session_token = NULL;
-    int session_token_len = 0;
+    size_t session_token_len = 0;
 
     if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ls", &php_options, &session_token, &session_token_len) ==
         FAILURE) {
@@ -188,7 +187,7 @@ PHP_FUNCTION(aws_crt_credentials_options_set_session_token) {
     }
 
     aws_crt_credentials_options *options = (void *)php_options;
-    aws_crt_credentials_options_set_session_token(options, (uint8_t *)session_token, (size_t)session_token_len);
+    aws_crt_credentials_options_set_session_token(options, (uint8_t *)session_token, session_token_len);
 }
 
 PHP_FUNCTION(aws_crt_credentials_options_set_expiration_timepoint_seconds) {
@@ -256,7 +255,7 @@ PHP_FUNCTION(aws_crt_credentials_provider_static_options_release) {
 PHP_FUNCTION(aws_crt_credentials_provider_static_options_set_access_key_id) {
     zend_ulong php_options = 0;
     const char *access_key_id = NULL;
-    int access_key_id_len = 0;
+    size_t access_key_id_len = 0;
 
     if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ls", &php_options, &access_key_id, &access_key_id_len) ==
         FAILURE) {
@@ -264,14 +263,13 @@ PHP_FUNCTION(aws_crt_credentials_provider_static_options_set_access_key_id) {
     }
 
     aws_crt_credentials_provider_static_options *options = (void *)php_options;
-    aws_crt_credentials_provider_static_options_set_access_key_id(
-        options, (uint8_t *)access_key_id, (size_t)access_key_id_len);
+    aws_crt_credentials_provider_static_options_set_access_key_id(options, (uint8_t *)access_key_id, access_key_id_len);
 }
 
 PHP_FUNCTION(aws_crt_credentials_provider_static_options_set_secret_access_key) {
     zend_ulong php_options = 0;
     const char *secret_access_key = NULL;
-    int secret_access_key_len = 0;
+    size_t secret_access_key_len = 0;
 
     if (zend_parse_parameters(
             ZEND_NUM_ARGS() TSRMLS_CC, "ls", &php_options, &secret_access_key, &secret_access_key_len) == FAILURE) {
@@ -280,13 +278,13 @@ PHP_FUNCTION(aws_crt_credentials_provider_static_options_set_secret_access_key) 
 
     aws_crt_credentials_provider_static_options *options = (void *)php_options;
     aws_crt_credentials_provider_static_options_set_secret_access_key(
-        options, (uint8_t *)secret_access_key, (size_t)secret_access_key_len);
+        options, (uint8_t *)secret_access_key, secret_access_key_len);
 }
 
 PHP_FUNCTION(aws_crt_credentials_provider_static_options_set_session_token) {
     zend_ulong php_options = 0;
     const char *session_token = NULL;
-    int session_token_len = 0;
+    size_t session_token_len = 0;
 
     if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "ls", &php_options, &session_token, &session_token_len) ==
         FAILURE) {
@@ -294,8 +292,7 @@ PHP_FUNCTION(aws_crt_credentials_provider_static_options_set_session_token) {
     }
 
     aws_crt_credentials_provider_static_options *options = (void *)php_options;
-    aws_crt_credentials_provider_static_options_set_session_token(
-        options, (uint8_t *)session_token, (size_t)session_token_len);
+    aws_crt_credentials_provider_static_options_set_session_token(options, (uint8_t *)session_token, session_token_len);
 }
 
 PHP_FUNCTION(aws_crt_credentials_provider_static_new) {
